move email service to common folder
