+++ src/modules/auth/auth.module.ts
@@ -12,7 +12,7 @@ import { RecoverPasswordService } from './services/recover-password.service';
 import { ResendAccountTokenEmailService } from './services/resend-account-token.service';
 import { UserRepository } from '../user/repository/user.repository';
 import { RedisCacheService } from '../../common/redis/redis-cache.service';
-import { EmailService } from '../user/services/email.service';
+import { EmailService } from '../../common/services/email.service';
 
 @Module({
   imports: [
