+++ src/modules/user/services/update-user.service.ts
@@ -7,7 +7,7 @@ import {
 } from '../interfaces/repository.interface';
 import { IUser, ISecurityData } from '../interfaces/user.interface';
 import { UpdateUserDto } from '../dto/update-user.dto';
-import { EmailService } from './email.service';
+import { EmailService } from '../../../common/services/email.service';
 import { Channel, User } from '@prisma/client';
 import { PasswordService } from './password.service';
 import {
