+++ src/modules/user/test/user.service.spec.ts
@@ -17,7 +17,7 @@ import {
   MockUserData,
 } from './mocks/user.mock';
 import { PasswordService } from '../services/password.service';
-import { EmailService } from '../services/email.service';
+import { EmailService } from '../../../common/services/email.service';
 import { RedisCacheService } from '../../../common/redis/redis-cache.service';
 import { Channel } from '@prisma/client';
 import { MailerService } from '@nestjs-modules/mailer';
