+++ src/modules/user/user.module.ts
@@ -8,7 +8,7 @@ import { UserController } from './user.controller';
 import { DeleteUserService } from './services/delete-user.service';
 import { GetUserByFilterService } from './services/user-by-filter.service';
 import { PasswordService } from './services/password.service';
-import { EmailService } from './services/email.service';
+import { EmailService } from '../../common/services/email.service';
 import { RedisCacheService } from '../../common/redis/redis-cache.service';
 
 @Module({
